clean up extracted files if they still exist

+++ app/jobs/process_uploaded_file_job.rb
@@ -56,11 +56,11 @@ class ProcessUploadedFileJob < ApplicationJob
           reader.extract(entry, Archive::EXTRACT_SECURE, destination: tmpdir.to_s)
           model.model_files.create(filename: filename, attachment: File.open(entry.pathname))
           # Clean up file
-          File.delete(entry.pathname)
+          File.delete(entry.pathname) if File.exist?(entry.pathname)
         end
       end
       # Clean up temp folder
-      Dir.rmdir(tmpdir)
+      Dir.rmdir(tmpdir) if Dir.empty?(tmpdir)
     end
   end
